+++ front/src/components/App.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Yahoo from './Yahoo';
-import Main from './Main';
+import HomePage from './Homepage';
 
 const App = () => {
   return (
     <div className="App">
       <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Main />} />
+        <Route path="/" element={<HomePage />} />
         <Route path="/news_yahoo" element={<Yahoo />} />
       </Routes>
       </BrowserRouter>
